Refactored to appease CodeClimate
