+++ lib/hyperion/aux/contracts.rb
