+++ lib/hyperion/aux/typho.rb
