+++ lib/hyperion/aux/util.rb
