+++ lib/hyperion/aux/version.rb
@@ -1,0 +1,3 @@
+class Hyperion
+  VERSION = '0.0.39'
+end
