--- lib/hyperion/version.rb
@@ -1,3 +1,0 @@
-class Hyperion
-  VERSION = '0.0.38'
-end
